fix(smoke): one browser context for the page sweep, run new specs last
The Clerk dev instance rate-limits aggressively; a fresh context per page
visit tripped 429s that failed the pre-existing smoke tests. The sweep now
reuses one page with pacing, and the new specs sort after the original
smoke so it always runs on a cold budget.
